Fixes for non-admin users and Steam fix
This fixes the app to understand when it's been run by a non-admin and
the game is installed to an admin-only directory (like Program Files).
It also tries to create a steam_appid.txt file in the AC4 install
directory which is mandatory for Steam to not forcibly start the
singleplayer executable.